requirements.txt
discord-py-slash-commands
discord.py
SQLAlchemy
sqlalchemy-migrate